Move hashes and write summary to own crate
