+++ bins/server/src/backbone/backbone.rs
@@ -2,11 +2,11 @@ use crate::backbone::file_reader::FileReader;
 use crate::backbone::file_record::{FileRecord, GetReaderError};
 use crate::backbone::file_writer::FileWriter;
 use crate::backbone::file_writer_guard::FileWriterGuard;
-use crate::backbone::WriteSummary;
 use crate::backends::{BackendCommand, BackendRegistry};
 use async_tempfile::TempFile;
 use axum::headers::ContentType;
 use axum::response::{IntoResponse, Response};
+use file_distribution::WriteSummary;
 use hyper::StatusCode;
 use rendezvous::RendezvousGuard;
 use shared_files::{SharedFileWriter, SharedTemporaryFile};
