+++ bins/server/src/backbone/file_record.rs
@@ -1,7 +1,7 @@
 use crate::backbone::backbone::BackboneCommand;
 use crate::backbone::file_writer_guard::WriteResult;
-use crate::backbone::WriteSummary;
 use axum::headers::ContentType;
+use file_distribution::WriteSummary;
 use shared_files::{SharedTemporaryFile, SharedTemporaryFileReader};
 use shortguid::ShortGuid;
 use std::sync::Arc;
