+++ bins/server/src/backbone/mod.rs
@@ -1,15 +1,12 @@
 mod backbone;
 mod file_accessor;
-mod file_hashes;
 mod file_reader;
 mod file_record;
 mod file_writer;
 mod file_writer_guard;
-mod hash;
 
 pub use backbone::Backbone;
 pub use file_accessor::{FileAccessor, FileAccessorBridge, FileAccessorError};
-pub use file_hashes::FileHashes;
 pub use file_reader::FileReader;
 pub use file_record::GetReaderError;
-pub use file_writer::{CompletionMode, WriteSummary};
+pub use file_writer::CompletionMode;
