+++ crates/file_distribution/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "file_distribution"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+md5 = "0.7.0"
+sha2 = "0.10.8"
+shortguid = "0.6.0"
+tokio = { version = "1.35.0", default-features = false, features = ["time"] }
