+++ crates/file_distribution/src/lib.rs
@@ -1,0 +1,6 @@
+mod file_hashes;
+pub mod hash;
+mod write_summary;
+
+pub use file_hashes::FileHashes;
+pub use write_summary::WriteSummary;
